refactor delete route for messages. get current user id to display only messages for that user. delete message by filtering by message id sent from front end. rename routes helper to helper routes

+++ app/routes/message_routes.py
@@ -5,7 +5,7 @@ from datetime import datetime, timedelta, timezone
 
 from app.models.user import User
 from app.models.message import Message
-
+y
 
 messages_bp = Blueprint("farewell messages", __name__, url_prefix="/farewell_messages")
 
@@ -54,21 +54,6 @@ def create_farewell_message():
         "msg": "Successfully created"
     }, 201
 
-# #Get all farewell messages created by the current signed in user
-# @messages_bp.route("", methods=['GET'])
-# @jwt_required(optional=True)
-# def handle_farewell_messages():
-#     message_query = request.args.get("messages")
-#     if message_query:
-#         messages = Message.query.filter_by(message=message_query)
-#     else:
-#         messages = Message.query.all()
-
-#     farewell_messages_response = []
-#     for message in messages :
-#         farewell_messages_response.append(message.to_dict())
-#     return jsonify(farewell_messages_response), 200
-
 #Get all messages from the current signed in user 
 @messages_bp.route("", methods=['GET'])
 @jwt_required()
@@ -87,15 +72,29 @@ def handle_farewell_messages():
         farewell_messages_response.append(message.to_dict())
     return jsonify(farewell_messages_response), 200
 
-    
-
 # DELETE ONE FAREWELL MESSAGE
-@messages_bp.route("/<message_id>/delete", methods=["DELETE"])
+@messages_bp.route("/<message_id>", methods=["DELETE"])
+@jwt_required()
 def delete_one_message(message_id):
-    message_to_delete = get_valid_item_by_id(Message, message_id)
+    current_user_email = get_jwt_identity()
+    try:
+        current_user = User.query.filter_by(email=current_user_email).first()
+    except Exception as e:
+        print("ERROR", str(e))
+        return {"Error": "An error ocurred when retriveing current user"}
+    
+    message_to_delete = Message.query.filter_by(user_id=current_user.id, id=message_id).first()
+    print("Query message:", message_to_delete)
+    
+    if message_to_delete is None:
+        return {"Error": "Message not found"}, 404
 
-    db.session.delete(message_to_delete)
-    db.session.commit()
-
-    return f"Message {message_to_delete} is deleted!", 200
+    try:
+        db.session.delete(message_to_delete)
+        db.session.commit()
+        return f"Message {message_id} is deleted!", 200
+    except Exception as e:
+        print("ERROR", str(e))
+        db.session.rollback()
+        return {"Error": "An error occurred while deleting the message"}, 500
